Implemented OR Rows
Changed the JSON output, adding functionality to the requirements parser to deal with requirements where one of a set of groups can fulfill a requirement.

Also added a comment to get addEnrollments to stop sending a wall of error messages when called; currently commented out so that THIS functionality can be effectively tested, but should probably be uncommented before pushing to development branch.